Moved Ship class into shared lib
