+++ client/index.html
@@ -5,8 +5,8 @@
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=1">
 
-<!--         <link href='https://fonts.googleapis.com/css?family=Aldrich' rel='stylesheet' type='text/css'>
-        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css"> -->
+        <link href='https://fonts.googleapis.com/css?family=Aldrich' rel='stylesheet' type='text/css'>
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css">
         <link rel='stylesheet' href='css/normalize.css'/>
         <link rel='stylesheet' href='css/main.css'/>
     </head>
@@ -59,7 +59,7 @@
         <script type='text/javascript' src='js/planet.js'></script>
         <script type='text/javascript' src='js/player.js'></script>
         <script type='text/javascript' src='js/moon.js'></script>
-        <script type='text/javascript' src='js/ship.js'></script>
+        <script type='text/javascript' src='js/lib/ship.js'></script>
         <script type='text/javascript' src='js/mouse.js'></script>
         <script type='text/javascript' src='js/main.js'></script>
         <script type='text/javascript' src='js/connection.js'></script>
