+++ lib/ship.js
@@ -1,6 +1,9 @@
+"use strict";
+
 // TODO: Constructor ID.
-var Ship = function(id, name, plot, speed) {
+var Ship = function(id, owner, name, plot, speed) {
     this.name = name;
+    this.owner = owner;
     this.id = id;
     this.size = 1 / 1000;
     this.speed = speed / 3600 || 1000 / 3600; // Defaults to 1000kmph
@@ -15,7 +18,11 @@ var Ship = function(id, name, plot, speed) {
     }
 
     return this;
-};
+}
+
+Ship.prototype.toString = function() {
+    return "SHIP: " + this.name + " [" + this.plot.x + "/" + this.plot.y + "]";
+}
 
 Ship.prototype.scaleShape = function(size) {
     var shapeCount = this.shape.length,
@@ -139,5 +146,6 @@ Ship.prototype.draw = function() {
         context.fillStyle = 'rgba(150, 255, 200, 0.3)';
         context.fillText(this.name, shipPosition.x + (size/2) - (labelWidth/2), shipPosition.y - this.size*2*game.scale - 2);
     }
-
 };
+
+if (typeof exports !== 'undefined') module.exports = Ship;
